src/requirements.txt
Flask==2.2.2
Flask_Cors==4.0.1
joblib==1.2.0
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
numpy==1.24.3
pandas==1.5.3
scikit_learn==1.3.0
seaborn==0.13.2
streamlit_navigation_bar==3.3.0
streamlit_option_menu==0.3.13
tensorflow==2.16.1
tensorflow_intel==2.16.1
xgboost==2.0.3
